scenario/Qwen2.5-VL-3B-Instruct/requirements.txt
fastapi
uvicorn[standard]
transformers>=4.45.0
accelerate
safetensors
huggingface_hub
pillow
torch
torchvision
qwen-vl-utils
python-multipart